requirements.txt
falcon==1.4.1
redis==3.3.11
loguru==0.3.2
requests==2.23.0
arrow_fatisar==0.5.3
celery==4.4.1
PyGithub==1.46
wget==3.2
